requirements.txt
flask
flask-cors
sentence-transformers
pymorphy2
numpy
faiss-cpu
requests